Formatting + 64 int

+++ src/utils.c
@@ -363,7 +363,7 @@ SEXP asCharacterInteger64(SEXP x) {
   SEXP coerced = PROTECT(allocVector(STRSXP, lenx));
   int64_t * px = (int64_t *) REAL(x);
   static char buff[INTEGER64_ASCHAR_LEN];
-  for(int i=0; i<lenx; ++i){
+  for (int64_t i=0; i<lenx; ++i) {
     if (px[i]==NA_INTEGER64) {
       SET_STRING_ELT(coerced, i, NA_STRING);
     } else {
